+++ FlashCap/FrameProcessors/IgnoreDroppingProcessor.cs
@@ -26,7 +26,7 @@ namespace FlashCap.FrameProcessors
         public override sealed void OnFrameArrived(
             CaptureDevice captureDevice,
             IntPtr pData, int size,
-            double timestampMicroseconds, long frameIndex)
+            long timestampMicroseconds, long frameIndex)
         {
             if (Interlocked.Increment(ref isin) == 1)
             {
